Use prebuild-config memfs mock
